Remove broken sibling selector optimization
Blink r154053 landed with a broken optimization for nth-child that
this patch removes.

Background: nth-child was formerly O(1) and nth-last-child was O(N^2),
but r154053 reversed these making the more common nth-child slower.
It included an optimization to make both O(1), but the optimization
was broken in one of the merges.

I will re-land the working optimization in another patch, and with
a magnitude perf test Ojan, eseidel, and myself wrote to ensure it
works.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/18147020

git-svn-id: svn://svn.chromium.org/blink/trunk@154072 bbb929c8-8fbe-4397-9dbb-9b2b20218538