Manage the lifetime of DOMTimers with OwnPtr.
The lifetime of DOMTimers are currently managed manually with new and delete.
This change wraps the DOMTimers with OwnPtr, and lets ScriptExecutionContext
own them.

This is a refactoring step that will be required to fix bug 230705. This is
a pure refactoring, and thus user-visible behavior should not change.

BUG=230705
R=kbr@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18240005

git-svn-id: svn://svn.chromium.org/blink/trunk@154342 bbb929c8-8fbe-4397-9dbb-9b2b20218538