Do not merge zero-width whitespace runs to other runs in complex text rendering.
This is a follow-up for r150623. Not all the "space" characters were handled there.

The rebaselining in this patch reverts http://trac.webkit.org/changeset/96602,
in which the text width in the expectation are shorten reflecting
the current bug that parenthesis after a rtl-ish mark disappears.

BUG=260102

Review URL: https://chromiumcodereview.appspot.com/19476003

git-svn-id: svn://svn.chromium.org/blink/trunk@154364 bbb929c8-8fbe-4397-9dbb-9b2b20218538