Record a Custom Element type extension for the life of the element.
Previously a Custom Element's type extension (if any) was stored in a
specific registry. However a type extension should persist for the
life of the element, which may span multiple registration contexts.

This is in preparation for processing Custom Elements which move
between registration contexts.

BUG=255725

Review URL: https://chromiumcodereview.appspot.com/19495003

git-svn-id: svn://svn.chromium.org/blink/trunk@154368 bbb929c8-8fbe-4397-9dbb-9b2b20218538