Remove and eliminate redundant, duplicative Custom Element resolution code.
This adds a 'resolve' method to resolve elements that enter the
purview of a Custom Element registration context. Currently that
happens when a custom tag element is created or a type extension is
set.

This change is in anticipation of a third caller, when an element is
adopted into a document that has an active registration context.

BUG=255725

Review URL: https://chromiumcodereview.appspot.com/19518004

git-svn-id: svn://svn.chromium.org/blink/trunk@154381 bbb929c8-8fbe-4397-9dbb-9b2b20218538