Merge viewport units/calcs conversion flags into fixed/percent flags
In CSSPrimitiveValue::computeLength, we decide which conversions to
allow based on a template parameter which is a bitwise OR of several
flags. Since viewport units are a <length> type (as per css3 values),
we should be converting them whenever we can convert to a fixed length.
This patch removes the ViewportPercentageConversion flag and enables the
conversion whenever we enable a fixed length conversion.
Similarly, we should also convert calcs with both percentage and length
components (the CalculatedConversion flag) whenever both are allowed.

BUG=

Review URL: https://chromiumcodereview.appspot.com/19500003

git-svn-id: svn://svn.chromium.org/blink/trunk@154417 bbb929c8-8fbe-4397-9dbb-9b2b20218538