Introduce HTMLDimension
This change patches the frameset code to use a new dimension class
instead of badly reusing Length, specifically Length::Relative (which
shouldn't exist as relative or multi-lengths are not a CSS concept).

The new class corresponds to a dimension as defined by HTML5 in
section 2.4.4.6 "list of dimensions". The HTML5 dimension is much
narrower than Length: the former covers 3 units (absolute, relative
and percentage), the latter covers CSS <length> (which is a lot of
dimensions) and <percentage>.

After this change, the only user of Length::Relative is the
border-image-width code.

BUG=259107

Review URL: https://chromiumcodereview.appspot.com/18986013

git-svn-id: svn://svn.chromium.org/blink/trunk@154554 bbb929c8-8fbe-4397-9dbb-9b2b20218538