When linting lines, never overwrite an old line.
Instead, always merge lines. This way, we always compare against all
previous lines when seeing if there are conflicts.

If there are 3 lines the apply to a given test that all have the
same specificity, the old code would only compare the first two for
conflicts and then the last two. It would never compare the first
to the last since the second one would overwrite the first before
we got to the last.

BUG=259124
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/19036002

git-svn-id: svn://svn.chromium.org/blink/trunk@154598 bbb929c8-8fbe-4397-9dbb-9b2b20218538