Remove ScopedPersistent::Is(Function|Object|String).
They use ScopedPersistent::deprecatedGet which we want to remove.

The callers used these functions to do sanity checks, and the "normal case" was
that the Value in question is indeed (Function|Object|String).

The callers have been migrated like this:
- String conversion & ScriptPreprocessor: we anyway created a local handle right
after the check, so create one a bit earlier.
- Other callers of Is(Function|Object): create a new Local handle and do the
check on that. The callers (MediaQueryListListener, InjectedScript.*) shouldn't
be so critical to performance that we cannot afford this.

BUG=236290

Review URL: https://chromiumcodereview.appspot.com/19772002

git-svn-id: svn://svn.chromium.org/blink/trunk@154639 bbb929c8-8fbe-4397-9dbb-9b2b20218538